phase2: move history timeline maintenance out of rewrite path
