+++ packages/engine/src/sql/history/mod.rs
@@ -1,1 +1,3 @@
+pub(crate) mod maintenance;
+pub(crate) mod requests;
 pub(crate) mod requirements;
